requirements.txt
PySide6>=6.0.0
witrnhid==1.0.0
hidapi>=1.0.0
psutil>=5.9.0